05. 選課輔助系統/requirements.txt
botbuilder-integration-aiohttp>=4.11.0
aiohttp
botbuilder-core>=4.7.0
botbuilder-dialogs>=4.7.0
botbuilder-ai>=4.7.0
botbuilder-azure>=4.7.0
pandas